Grit and Cure issue is solved Grit was a problem in the explosion ability, didn't send a target, fixed now. Cure tried to getComponent a character, solved it with getComponent Unit. AI does now account for movement when using abilities. judgeAbility now uses float instead of int. Explosion had the same issue as Cure, solved it with getComponent Unit

+++ Assets/Src/Abilities/AIAbility.cs
@@ -10,7 +10,7 @@
 using System;
 public interface AIAbility
 {
-	int judgeAbility(Unit user, Tile move, out Tile target);
+	float judgeAbility(Unit user, Tile move, out Tile target);
 	Ability getAbility
 	{
 		get;
